requirements.txt
# Core Agentic Framework
langgraph>=0.2.0,<0.3.0
langchain>=0.3.0,<0.4.0
langchain-ollama>=0.1.0
langchain-community>=0.3.0

# Data Processing & Forecasting
pandas>=2.2.0,<2.3.0
numpy>=1.26.0,<1.27.0
prophet>=1.1.0,<1.2.0
statsmodels>=0.14.0,<0.15.0
scikit-learn>=1.5.0,<1.6.0

# Database
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0,<3.0.0

# Visualization
matplotlib>=3.9.0,<4.0.0
plotly>=5.24.0,<6.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
loguru>=0.7.0,<0.8.0
schedule>=1.2.0,<1.3.0

# HTTP Requests (for potential IESO API calls)
requests>=2.31.0,<3.0.0
httpx>=0.27.0,<1.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=5.0.0,<6.0.0

# Code Quality
black>=24.0.0,<25.0.0
ruff>=0.6.0,<1.0.0

# Type Hints
typing-extensions>=4.12.0,<5.0.0
